refactor: Refactored some elements in projects and themes pages to allow for more server side rendering
