+++ public/locales/en/translation.json
@@ -92,8 +92,7 @@
 		"experiencebutton": "My Experience"
 	},
 	"Themes": {
-		"title": "Color Theme for This Website",
-		"accessDenied": "Access denied. Please enable developer mode to access this page."
+		"title": "Color Theme for This Website"
 	},
 	"Projects": {
 		"title": "Projects",
@@ -120,5 +119,8 @@
 				"projectID": "gr2313"
 			}
 		]
+	},
+	"DevMode": {
+		"accessDenied": "Access denied. Please enable developer mode to access this page."
 	}
 }
